blockd: fix length of timeout int passed to ioctl
AUTOFS_IOC_SETTIMEOUT expects a pointer to an 'unsigned long' which
will result in out-of-bounds access when passing a pointer to an 'int'.
Change type of timeout to 'unsigned long'.

+++ blockd.c
@@ -574,7 +574,7 @@ struct uloop_timeout autofs_expire_timer = {
 
 static int autofs_mount(void)
 {
-	int autofs_timeout = AUTOFS_TIMEOUT;
+	unsigned long autofs_timeout = AUTOFS_TIMEOUT;
 	int kproto_version;
 	int pipefd[2];
 	char source[64];
